Integrate @web5/identity-agent to create identity on app launch
This PR pulls in the `@web5/identity-agent` package, and uses it to create an identity when the user launches the app.

Overview of the Identity Agent: https://hackmd.io/ShKU_Q9aQC6C96v7SaXl3A?view#Overview

The majority of this work was polyfills, and getting those right. For the time being, I've added those polyfills directly to this repo, as I'm not sure how to publish a new version of `web5-react-native-polyfills`. We can move these polyfills to the `web5-react-native-polyfills` repo once Tim's back from paternity leave!