Fix crash on Android
On Android, `getBufferJava` is called to get a `jbyteArray` of the backing bytes for the Blob. However, as we've seen on iOS, these bytes m
ay not be available in such quick succession as Web5 is creating them, then immediately accessing them. This causes the backing bytes to not
be available on Android, and an exception is throw from within the `getBufferJava` call.

The existing react-native-blob-jsi-helper package does not seem to account for this, and assumes it works every time. This isn't the case.
So, in this patch, I detect that there's an error, and throw a new error. This allows our JS code to detect that the bytes haven't been read
yet, and can schedule a microsleep to get it again in the future.