Integrate @web5/identity-agent to create identity on app launch (#62)
* Integrate @web5/identity-agent to create identity on app launch

This PR pulls in the `@web5/identity-agent` package, and uses it to create an identity when the user launches the app.

Overview of the Identity Agent: https://hackmd.io/ShKU_Q9aQC6C96v7SaXl3A?view#Overview

The majority of this work was polyfills, and getting those right. For the time being, I've added those polyfills directly to this repo, as I'm not sure how to publish a new version of `web5-react-native-polyfills`. We can move these polyfills to the `web5-react-native-polyfills` repo once Tim's back from paternity leave!

* Fix crash on Android

On Android, `getBufferJava` is called to get a `jbyteArray` of the backing bytes for the Blob. However, as we've seen on iOS, these bytes m
ay not be available in such quick succession as Web5 is creating them, then immediately accessing them. This causes the backing bytes to not
be available on Android, and an exception is throw from within the `getBufferJava` call.

The existing react-native-blob-jsi-helper package does not seem to account for this, and assumes it works every time. This isn't the case.
So, in this patch, I detect that there's an error, and throw a new error. This allows our JS code to detect that the bytes haven't been read
yet, and can schedule a microsleep to get it again in the future.

* Remove var

* pr feedback

* Add a TODO
* Don't cast global
* Remove async keyword from async function